Refactored into OO design.
1) DictionaryList: class that conforms to the getWord method interface.
The dictionary is contained in a list. At some point I'll make another
that just seeks through the file and because of the common interface
they could be swapped freely.
2) RandomDict: A subclass of DictionaryList that returns a random word.
It's implemented as a generator, so just call the next() method.
3) PasswordGen: This gets words from RandomDict and makes a password.
Also returns password information. Implemented as a generator, so call
next() to get more passwords.