Added two experimental scripts to test a potential algorithm.
1) dice_rolls: simluates rolling a crap ton of dice, and picking the
highest 5. This is done for many trials, and a tally of how often each
die gets picked is kept. This seems to indicate that each die has an
equal chance of being picked. That sounds kind of obvious, but wanted to test it
and make sure.
2) rand_proto: This is an experimental implementation of the dice
rolling algorithm applied to word selection. Each word is assigned a
random value. The words with the lowest value get picked. Ties are
broken randomly. The nice thing is that only a small faction of the
words need to be kept in memory, so it's much more efficient in that
sense.